% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPreprocess.R
\name{ReadGenomicPLMM}
\alias{ReadGenomicPLMM}
\title{Multi-kernel linear mixed model with adaptive lasso for prediction analysis on high-dimensional multi-omics data}
\usage{
ReadGenomicPLMM(bed, bim, fam, na.strings = c("0", "-9"), phenofile,
  trainID, annotation = NULL, Y = NULL, X = NULL, Kinship = NULL,
  kernels = c("linear", "poly2"), AllRegions = 0)
}
\arguments{
\item{bed}{The name of the file containing the packed binary SNP genotype data. It should have the extension .bed; if it doesn't, then this extension will be appended}

\item{bim}{The file containing the SNP descriptions}

\item{fam}{The file containing subject (and, possibly, family) identifiers. This is basically a tab-delimited "pedfile"}

\item{na.strings}{Strings in .bam and .fam files to be recoded as NA}

\item{phenofile}{The file containing phenotype information. If provided, will ignore the affected values from the .fam file. The file should contain two columns: pheno: phenotypic values. ID: subject ID for each indidivuals.}

\item{trainID}{Subject ID for the training individuals.}

\item{annotation}{A data frame providing information to cut the genomes. It has at least four columns: gene, chr, start, end. The gene column must be unique. The other three columns contain information about chromosome, start and end of each regomic region. If Kinship is not provided, then annotation must be provided.}

\item{Y}{A vector of phenotypes with each name being subject ID. If provided, will ignore \code{phenofile} and \code{fam}.}

\item{X}{A matrix of demographic variables, should be of the same order as Y (i.e. \code{rownames(X)=names(Y)}). The intercept column is not needed.}

\item{Kinship}{A list of the pre-calculated genomic similarity matrix, with each element being a genomic region. The row and column names for each element should be the subjectID. If provided, will ignore \code{bed} and \code{annotation}.}

\item{kernels}{A vector of strings only taking two values: linear (linear kernel) and poly2 (polynomial kernel with 2 df).}

\item{AllRegion}{A bool to indicate whether GSM should be calculated from the entire genome. Default=0.}
}
\value{
A list that contains Y, Kinship and train.index, all of which are needed for KLMMAL algorithm(i.e. the function \code{KLMMALPred}).
}
\description{
\code{ReadGenomicPLMM} prepares the data for analyses
}
\details{
This function is used to pre-process the data used by the Multi-Kernel Linear Mixed Model with Adaptive Lasso method.
}
\examples{
## Using the data with pre-calculated GSMs ##
load(system.file("extdata", "Data.rda", package = "KLMMAL"))

## read both genotype and phenotype files from the disk, and calculate the GSMs using linear kernels ##
trainID=read.table(system.file("extdata", "TrainID.txt", package = "KLMMAL"));
annotation=read.table(system.file("extdata", "Annotation.txt", package = "KLMMAL"),header=T);
Data=ReadGenomicPLMM(bed=system.file("extdata", "testing.bed", package = "KLMMAL"),trainID=trainID[,1], annotation = annotation, phenofile=system.file("extdata", "pheno.txt", package = "KLMMAL"), kernels = c("linear"), AllRegions = 0);

## read genotype from the disk and calculate the GSMs using linear and poly2 kernels##
pheno=read.table(system.file("extdata", "pheno.txt", package = "KLMMAL"),header=T);
Y=pheno$pheno;names(Y)=pheno$ID;
trainID=read.table(system.file("extdata", "TrainID.txt", package = "KLMMAL"));
annotation=read.table(system.file("extdata", "Annotation.txt", package = "KLMMAL"),header=T);
Data=ReadGenomicPLMM(bed=system.file("extdata", "testing.bed", package = "KLMMAL"),trainID=trainID[,1], annotation = annotation, Y = Y, kernels = c("linear", "poly2"), AllRegions = 0);
}
